+++ src/engine/shapes/Plane.h
@@ -30,9 +30,9 @@
 class Plane : public Shape {
 protected:
     Vector3 normal;
-    double distance;
+    float distance;
 public :
-    Plane(const Vector3& n, const double d);
+    Plane(const Vector3& n, const float d);
     Plane(const Vector3& P, const Vector3& n);
     Plane(const Vector3& C, const Vector3& v1, const Vector3& v2);
     GPU::GPUShapeData toGPU() const override;
